requirements.txt
pandas 
torch 
Pillow
torchvision 
progress
numpy
matplotlib 
path
scikit-image 
Rtree
opencv_python 
scipy 
tqdm